Using ELF file for additional symbols and data
`static` symbols are missing from the map file but are present in the
elf file. Since the ELF file contains both these and the final binary
data, this is now the second argument to `evaluate` rather than the BIN
file. The map file is created when building the `elf` file so this
should not be a huge concern.

Renamed to `mipsmatch`.