Initial support for PSP Map Files
PSP contains several `__fill__` segments which cause issues when trying
to create a fingerprint for a particular file. This also interrupts the
normal segment to multiple function symbol layout of the PS1 map files.

This has been resolved but the resulting fingerprints are not matching
even the files they were generated from.