Check for SHF_EXECINSTR When Filtering ELF Sections
Previously MetroWorks headers were exluded when looking for `PROGBITS`
by excluding the section by name. Binaries compiled with GCC which had
headers were not being considered. GCC uses a different header name, so
these were not properly excluded.

Now the `SHF_EXECINSTR` flag is checked to determine which `PROGBITS`
section contains program code. This should be all that's necessary, but
an additional check for the MW and GCC headers has been left in for now.