Support for Matching `.rodata` segments
Fingerprints now include a description and size for any `.rodata`
section which might be part of a particular oject. There are currently
four matching strategies for these using jump tables to determine their
location.

In the best case, `.rodata` consists solely of jump tables. This is easy
to identify when scanning.

The other implemented case is when jump table data can be found at the
end of a `.rodata` segment.

Fingerprint configuration exists for jump tables at the beginning of a
segment as well as surrounding non-jump table data, but those cannot be
scanned for yet.

Cases where `.rodata` contains only data and no jump tables or const
data comes before and after jump tables are not supported.

To search for `.rodata` when using `scan`, the `--vram-offset` option
must be used with the location the overlay will be loaded into VRAM.

Despite checking processor family, instructions were only ever read in
as little endian. This has been fixed but will result in different
fingerprints for big endian systems. Since there is no real support for
those right now anyway, this does not affect any users.