Moved RK Search into Hasher
Moves the remaining Rabin-Karp specific code into `rk.rs` to abstract it
from fingerprinting and scanning.

The updated find function borrows the binary slice and lazily normalizes
instructions. This avoids copying the bin array into an u32 array, but
there is enough that needs it in that form that it might not be
worthwhile.

Added tests for scanning and unit tests for RK.